mid: 0000507
msg: Cambios en la presentación del listado de las fichas informativas asignadas a un procedimiento.

git-svn-id: https://svn.code.sf.net/p/rolsac/code/rolsac-core/branches/rolsac-1.3@1018 58ef2255-6f3f-491f-b93b-e266552ec6fa

+++ moduls/model/src/org/ibit/rol/sac/model/dto/SeccionFichaDTO.java
@@ -10,15 +10,15 @@ public class SeccionFichaDTO extends IdNomDTO implements Serializable {
 	// Cuidado: esta variable no equivale al size() de listaFichas.
 	// Se usa cuando no queremos inicializar la lista de fichas por eficiencia y, aún así,
 	// conocer el número de fichas relacionadas con una UA y una sección.
-	private Long numFichas;
+	private int numFichas;
 	
 
-	public Long getNumFichas() {
+	public int getNumFichas() {
 		return numFichas;
 	}
 	
 
-	public void setNumFichas(Long numFichas) {
+	public void setNumFichas(int numFichas) {
 		this.numFichas = numFichas;
 	}
